test_client.pyを削除しました
